TowerBuildMenu with Pause Menu UI
