--- Assets/Scripts/CustomPauseMenu.cs
@@ -1,45 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CustomPauseMenu : MonoBehaviour
-{
-
-    public static bool GamePaused = false;
-    public GameObject PauseMenuUI;
-
-    // Update is called once per frame
-    void Update()
-    {
-        if (Input.GetKeyDown(KeyCode.Escape))
-        {
-            if (GamePaused)
-            {
-                Resume();
-            }
-            else
-            {
-                Pause();
-            }
-        }
-    }
-   public void Resume()
-    {
-        {
-            PauseMenuUI.SetActive(false);
-            Time.timeScale = 1f;
-            GamePaused = false;
-        }
-
-    }
-    void Pause()
-    {
-        PauseMenuUI.SetActive(true);
-        Time.timeScale = 0f;
-        GamePaused = true;
-    }
-    public void Quit()
-    {
-        Application.Quit();
-    }
-}
